+++ CustomExceptionHandling/DuplicateProductException.java
@@ -1,0 +1,6 @@
+package CustomExceptionHandling;
+public class DuplicateProductException extends Exception {
+    public DuplicateProductException(String message) {
+      super(message);
+    }
+  }
