+++ CustomExceptionHandling/InsufficientQuantityException.java
@@ -1,0 +1,6 @@
+package CustomExceptionHandling;
+public class InsufficientQuantityException extends Exception {
+    public InsufficientQuantityException(String message) {
+      super(message);
+    }
+  }
