+++ CustomExceptionHandling/InvalidProductDataException.java
@@ -1,0 +1,6 @@
+package CustomExceptionHandling;
+public class InvalidProductDataException extends Exception {
+    public InvalidProductDataException(String message) {
+      super(message);
+    }
+  }
