+++ CustomExceptionHandling/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package CustomExceptionHandling;
+
+public class ProductNotFoundException extends Exception {
+    public ProductNotFoundException(String message) {
+      super(message);
+    }
+  }
