--- DuplicateProductException.java
@@ -1,5 +1,0 @@
-public class DuplicateProductException extends Exception {
-    public DuplicateProductException(String message) {
-      super(message);
-    }
-  }
