--- InsufficientQuantityException.java
@@ -1,5 +1,0 @@
-public class InsufficientQuantityException extends Exception {
-    public InsufficientQuantityException(String message) {
-      super(message);
-    }
-  }
