--- InvalidProductDataException.java
@@ -1,5 +1,0 @@
-public class InvalidProductDataException extends Exception {
-    public InvalidProductDataException(String message) {
-      super(message);
-    }
-  }
