+++ InventoryManagementSystem.java
@@ -1,4 +1,5 @@
 import java.util.Scanner;
+import CustomExceptionHandling.*;
 
 public class InventoryManagementSystem {
     private ProductManager productManager;
