+++ Product.java
@@ -1,3 +1,5 @@
+import CustomExceptionHandling.InvalidProductDataException;
+
 public class Product {
   private int id;
   private String name;
