--- ProductNotFoundException.java
@@ -1,5 +1,0 @@
-public class ProductNotFoundException extends Exception {
-    public ProductNotFoundException(String message) {
-      super(message);
-    }
-  }
